Rename the API to v1
