+++ lib/octocatalog-diff/api/v1/catalog-diff.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module OctocatalogDiff
+  module API
+    module V1
+      # This class allows octocatalog-diff to be used to obtain the difference between
+      # two catalogs.
+      class CatalogDiff
+        #
+      end
+    end
+  end
+end
